ANIMATED BUTTON
This code creates an animated button with a 3D effect and a sliding color transition on hover. The button is designed to be responsive and will adapt to the size of the container it's in.

To use this code, simply copy and paste the HTML and CSS code into your project. You can customize the button by adjusting the CSS properties, such as the font size and background color.

README
To use this code:

Copy and paste the HTML and CSS code into your project.
Adjust the CSS properties to customize the button.
Enjoy your awesome animated button!